succesful move to mixin
for individual blog posts, I've moved the api call into a mixin.

Still plenty of organization to do, and of course the remaining API calls need to be converted, but at least i've got the idea in place.  Lots of fun learning about what exactly Axios is returning, and the proper way to return the Axios call from the mixin's method